Rename RawTargetNode to UnconfiguredTargetNode
Summary:
`RawTargetNode` is a node in an unconfigured target graph (more or less), it is loaded by `UnconfiguredBuildTargetView`, e. g. `RawTargetNodePipeline.getNodeJob` accepts `UnconfiguredBuildTargetView` parameter.

Current loading sequence is this:

`BuildFileManifest entry -> Map<String, Object> -> RawTargetNode -> TargetNode`.

So it would be more consistent to rename it to `UnconfiguredTargetNode` (and other classes like `RawTargetNodePipeline` to `UnconfiguredTargetNodePipeline`).

Note `UnconfiguredTargetNode` is a map to uncoerced objects, that should be changed.

This diff is just mechanical rename, nothing more.

Reviewed By: bobyangyf

shipit-source-id: e429073ad9