Kotlin friend_paths should also apply to the ABI jar
Summary:
Kotlin has the concept of "friend paths" (https://github.com/JetBrains/kotlin/blob/master/compiler/cli/cli-common/src/org/jetbrains/kotlin/cli/common/arguments/K2JVMCompilerArguments.kt#L276) which allow you to specify libraries for which you are able to access anything marked `internal`. These are commonly used to provide extra access for unit tests.

Now that we are building against ABIs (or will be shortly), we need to also pass the SourcePath for the ABI to the compiler so that it can also access the internal functions in the ABI.

Reviewed By: asp2insp

shipit-source-id: e2b056d581