ConfigurationRuleResolver.getRule now accepts rule type as a parameter
Summary:
We do not have a proper generic cycle depency detection in configuration rules.

This feature fix is a simple (and probably temporary) work around this limitation: rule type is checked before rule is created, and since configuration rules depend on rules of different types:

* platform -> constraint_value -> constraint_setting
* config_setting -> constraint_value -> constraint_setting

This way we effectively prevent cycles in configuration rules.

This change is needed to allow referencing configuration rules from configuration rules directly (as in happens in build rules), instead of referencing targets.

E. g. `ConstraintValueRule` can now reference `ConstraintSettingRule` (implemented further on the stack).

Which is in turn needed to prevent non-unique constraint settings in a set of constraint values.

Reviewed By: bobyangyf

shipit-source-id: 06adc8060c