No longer add recursive resources to buck project
Summary: Previously, buck project had the concept of `recursiveAssets` which were a lot of the assets needed to build and run the app in simulator -- like the launch images or app icons, e.g. These were only needed because of how Xcode was being used to build the app. Now that we're using buck to build and Xcode as a straight IDE, we can avoid this and reduce the amount of cruft generated by buck project.

Reviewed By: williamtwilson

shipit-source-id: 66af347a6f