Move generated files to GENERATED directory within a package
Summary:
Up until now, generated files were correctly place in a buck-out directory in Xcode representing their place on disk, as expected. However, given the highly nested nature of this directory, it became cumbersome to see easily and clearly which generated files were part of your buck target.

To ease this, this moves generated files out of the `buck-out` directory and into a simple generated directory under the package.

- All generated files in a package, regardless of a target, are put in the same folder under the package name; this is because there is no canonical way to determine the 'root' folder of a target (since that doesn't really exist).
- The naming scheme is `GENERATED_{package_name}` for consistency.

Reviewed By: williamtwilson

shipit-source-id: 6bfbb37dea