Do not store absolute paths in basePath, ensure they are relative to cellPath
Summary:
Previously we stored absolute paths in the 'basePath' component of BuildFileSpec objects. We also pass in the cellPath, so this doesn't seem to make much sense (I saw several comments that didn't get why we did this along the way).

So, don't pass in absolute paths, and relativize to the cell if we do.

Reviewed By: jtorkkola

shipit-source-id: 5495b90ee5