Print a more helpful error if relative target's base path does not exist in PWD, but exists in root cell
Summary:
Previously if a user was in subdir `foo` and ran `buck build bar:baz`, we would try to build `//bar:baz`
Now, we by default try to build `//foo/bar:baz`. If the user previously was running `buck build bar:baz`, they would now get an error saying `foo/bar` does not exist, which can be confusing.

If the full package path does not exist, but the package path of the non-modified arg /does/ exist at the root, explain to the user how to correct the issue. Otherwise, just print the old message

Reviewed By: joshkehn

shipit-source-id: 36c9ec4c74