Fix crash in Path.relativize during IntelliJ project creation
Summary:
JCL implementations of `Path.relativize`, namely `UnixPath`, no longer support `".."` in the base path. We rely on this behavior in IntelliJ project creation.

The fix is to add our own `relativizePath` method that does essentially what the Java 8 version did, but short circuit to `BuckUnixPath.relativize` when possible for better perf.

I've opted to not use this new logic in the existing `MorePaths.relativize` to avoid a potential perf hit when the Buck filesystem provider is not being used, as this code is likely slower than the JCL's code, which has access to private methods for constructing paths more efficiently.

Reviewed By: sbalabanov

shipit-source-id: 9032eb6411