Fix empty JAR test in JavaDocTest
Summary:
`ZipStep` doesn't actually create an empty ZIP file when there are 0 entries to add. It always adds an entry with an empty filename and no content. This appears to be unintentional, and happens because `FileVisitor::visitFile` also gets called for the root path.

In Java 8, when we query a ZIP file with `ZipArchive.getDirNames`, the underlying JCL ZIP file walker doesn't return empty filename entries, but in Java 11 it does, causing a `JavadocTest` failure on Java 11.

The fix is to create a truly empty ZIP file when we have no entries to add.

Reviewed By: philipjameson, bobyangyf

shipit-source-id: d102d8f269