Print out stdout and stderr from RE
Summary:
This is what I'd like to be able to land to make it so that we report stdout and stderr from RE actions back to the user's console. I think it's reasonable to make stdout a warning and stderr an error, though happy to hear suggestions. I like this because it should eventually mean that the unused dependencies reporting happens in the same way (with yellow warning text).

There are a few things that need fixing before this can happen:
1. Right now, stderr from RE is full of errors relating to fields of a Buildable needing to be final: https://fburl.com/diffusion/ekxjmlx2. We need to either fix these, or downgrade this to a lower level. I have seen two so far, one is https://fburl.com/diffusion/ncvfymz7 and the other is https://fburl.com/diffusion/pppbb4hr, which I am not sure how to fix.

2. Right now, all events from the console get logged to stderr: https://fburl.com/diffusion/z7wzibhq Is that expected? Could we log to stdout for events up to `WARN` level, and log to stderr for events with `ERROR` level? Then we can distinguish between the two when we get them back on the dev's machine.

Raj Yengisetty, Chris Hopman, please let me know the best way to proceed on these two points. Thanks!

Reviewed By: rajyengi

shipit-source-id: 9c08347236