Normalize output path for genrules
Summary: The "out" parameter of a genrule is not necessarily a normalized path. In fact, it is not uncommon for genrules to use "." as the out parameter to refer to an entire directory of generated files. Filesystem system calls generally require a normalized path, so this commit explicitly normalizes the genrule's output path.

Reviewed By: mykola-semko

shipit-source-id: 34654f9c55