Fix Apple project test failures due to Path.relativize now behaving sanely
Summary:
`Path.relativize` behaves differently in Java 11 vs. Java 8 in two ways: (1) it normalizes both paths first, and (2) it doesn't support `.` or `..` in the base path. Java 8 returns questionable results from `Path.relativize` for unnormalized inputs, e.g. relativizing `child` against `.` results in `../child`, as `.` is treated like a normal path segment. This is why the `MorePaths.relativize` helper exists, but it isn't used everywhere.

Some of the tests relied on this broken behavior, causing failures in Java 11, where `Path.relativize` behaves more sanely. I've updated these tests to use `..` as the base instead of `.`, which is a bit saner given the expected relativized paths in the test (prefixed by `../`) , and changed the product code to use the new `MorePaths` method I added in a previous change, which makes the behavior consistent between Java 8 and 11. I'm not inclined to comb through the project generation code to properly fix code using and/or relying on the bogus `relativize` behavior, as this seems risky, and out of scope.

Reviewed By: stepancheg

shipit-source-id: 9f064d8b62