Basic version of configuration detector
Summary:
This version only detects targets by package prefix.

In this version detector if configured in `.buckconfig` as

```
[parser]
    target_platform_detector_spec = \
        target://foo/...->//:foo-p \
        target://bar/...->//:bar-p
```

This configuration means that that if:

* platform is not specified using `--target-platforms=` flag
* targets do not specify `default_target_platform`

then for all targets in `//foo/...` default target platform will be `//:foo-p`, and for all targets in `//bar/...` default target platform will be `//:bar-p`, and for all other targets target configuration will be "unconfigured".

I want to explicitly mention that **detector only applied to targets specified on command line**, dependencies of these targets inherit the configuration.

## Future development

* implement detection by host os
* implement detection by rule type
* implement detection of host platform
* implement nested rules (to implement conditions like "if prefix is //aaa/ and host operating system is linux")

Probably we want to:
* move detector configuration into `BUCK` files instead of `.buckconfig`
* **or** move it into `WORKSPACE` file which we don't yet have

Reviewed By: bobyangyf

shipit-source-id: eef98a25eb