Allow specifying target-platform in buckconfig
Summary:
* add `parser.target_platforms` config
* remove `build.host_platform` config

Previously `build.host_platform` config was used as a fallback when `--target-platforms=` is not specified. This option is not used, and it is removed.

`parser.target_platforms` option is added, and it is used as a fallback when `--target-platforms=` command line option is not specified.

This option should be used for transition only, but when migration to new platforms finished, this option should be removed.

Reviewed By: bobyangyf

shipit-source-id: 5f945a9174