Refactor UnusedDependenciesFinder to only use allowed inputs
Summary:
UnusedDependenciesFinder was using inputs that are not allowed at execution time, such as BuildRules. This refactors it to use just the build target strings and also the outputs of the rules that it depends upon.

There is a small behavioural change here which is that we only allow an "unused dep" to not be classed as unused if one of its exported deps is used, rather than checking the whole exported dependency chain. That seems like a reasonable change.

N.B. This is still "capturing" things somewhat, I'll change that in the next diff.

Reviewed By: cjhopman

shipit-source-id: eb1f407dd2