Tell users that they can use runtime_deps to avoid unused dependency checking
Summary: There are legit cases where a dependency is not required for compilation, but it is required in the deps (e.g. for reflection to occur). In this case, rather than removing the dep, you can tell Buck that it is a `runtime_dep`, and the unused dependency checker will ignore it.

Reviewed By: asp2insp

shipit-source-id: 5782353348