Add ability to parse label regex
Summary:
We would like to add the ability to group libraries by matching labels. In this diff, we extend the supported syntax in `BUCK` files, so that such a feature can be supported.

Particularly, we add the ability to add a 3rd element to the link group mapping which starts with `label:` and the string after the colon is a regex that will be used for matching.

Reviewed By: paulqi1990

shipit-source-id: de0f4abffa