Refactor method to support multiple targets per mapping
Summary:
Previously, `CxxLinkGroupMappingTarget` only represented a single `BuildTarget` which to apply grouping for. With the addition of label filtering, that's no longer the case.

In this diff, we prepare for the change in logic by extracting the code to allow for a single `CxxLinkGroupMappingTarget` to represent multiple `BuildTarget`s.

NOTE: No behavior change in this diff.

Reviewed By: paulqi1990

shipit-source-id: 66b1d395bd