Add grouping using label regex matching
Summary:
## Context

At present, link groups can be specified by listing build targets and traversal type (`tree` / `node`). While that's explicit and easier to read, it's not suitable for situations where the list of targets is determined in a more dynamic fashion without centralised control. In any production app, we need teh ability to combine a more static definition together with a more dynamic one.

## Label Filtering

To solve the above problem, we introduce the ability to match nodes for the purposes of grouping by filtering based on a regex which is matched against the labels of targets. For example:

  ("group_name", [("//Some:Target", "tree")])

This is the traditional spec which will group all transitive deps of `//Some:Target` under `group_name`. In this diff, we extend this functionality to be able to specify:

  ("group_name", [("//Some:Target", "tree", "label:some_regex")])

The semantics of the above are that any nodes, part of the transitive closure of the graph rooted at `//Some:Target`, which have an ancestor with a label that matches `some_regex` will belong to `group_name`.

For example, let's assume that there's a node `//Lib/A:B` in the transitive dep graph of `//Some:Target` which has a label `some_regex`. That means that `//Lib/A:B` and all of its transitive deps (regardless of whether they have a matching label or not), will become part of `group_name`.

Another way to think about it is that in the definition above, `//Some:Target` is replaced with all targets which have a matching label. That is:

  ("group_name", [("...all targets in closure of //Some:Target which have a label matching some_regex...", "tree", "label:some_regex")])

In production, `//Some:Target` would usually be replaced with a root placeholder library although that's not necessary.

The equivalent logic applies for `node` traversal.

## Design Choice

The particular design was chosen as it anchors the search within the transitive closure rooted at a particular node. If we omitted the node, we would need to traverse the *whole* dep graph in Buck. An alternative would have been to support a syntax similar to:

  ("group_name", [("label:some_regex", "tree")])

NOTE: The current design choice does not preclude the alternative in the future, so it can be extended further if deemed necessary.

shipit-source-id: a6fc05023d