Fix issue where header search paths were not correctly writing to buck-out
Summary: The issue here was  related to builds that were located in a different cell. The `HeaderSearchPaths` class was doing everything relative to the target cell, which meant that the path was being resolved incorrectly. To fix this, we need to change this to use the projectFileSystem for all target.

shipit-source-id: 3d766c7f53