Intern RuleBasedTargetConfiguration
Summary:
Practically we have only half a dozen of configurations in one build, so intern the object to save memory.

If we don't intern, we allocate about 100K object, which is probably not a big issue, but better safe than sorry.

`BuildTarget` is quite heavy object, it references `Path` which references filesystem, so use weak interning to avoid memory leaks.

Reviewed By: bobyangyf

shipit-source-id: 154cdb1c3e