Why not allow in place python binary creation.
Summary:
Buck requires symlinks now, has for a while, and this
restriction was added in 2015. Its not actually a restriction anymore.
However, the actual run_inplace.py.in file needs to be updated to
account for Windows specific isms, like, a lack of a real exec, and
wonky signal handling. Fix that, fix tests, things look good.

Reviewed By: stepancheg, philipjameson

shipit-source-id: bff325d57f