Parsing logic to properly handle -f for xcodebuildtool
Summary: Passing -f to buck is the standard way of filtering out tests to run. This is normally implemented by listing the tests and filtering them out according to the given filters, but xcodebuild doesn't have a built in way to list tests at all. Implementations of xctool that use xcodebuild directly can't list the tests, so this logic will transform a subset of the -f expressions into a form xcodebuild can understand.

shipit-source-id: fe37a95500