More compact graphviz output
Summary:
`query` and `audit actiongraph` have options to output in GraphViz DOT
format. Currently that output is very human-readable, but also very
redundant:

```
digraph action_graph {
  "//assets/android:agent.apk" [style=filled,color="#A7AD9D"];
  "//assets/android:native-exopackage-fakes.apk" [style=filled,color="#A7AD9D"];
  "//config:logging.properties" [style=filled,color="#A7AD9D"];
  "//programs:buck" [style=filled,color="#DFF4D6"];
  "//programs:buck" -> "//assets/android:agent.apk";
  "//programs:buck" -> "//assets/android:native-exopackage-fakes.apk";
  "//programs:buck" -> "//config:logging.properties";
```

This diff adds another output option that uses integer IDs for the nodes
and moves the names to the `label` attribute:

```
digraph action_graph {
  1 [style=filled,color="#A7AD9D",label="//assets/android:agent.apk"];
  2 [style=filled,color="#A7AD9D",label="//assets/android:native-exopackage-fakes.apk"];
  3 [style=filled,color="#A7AD9D",label="//config:logging.properties"];
  4 [style=filled,color="#DFF4D6",label="//programs:buck"];
  4 -> 1;
  4 -> 2;
  4 -> 3;
```

The rendered output is the same, but this second approach is drastically
smaller and faster for machines to parse large graphs.

Reviewed By: jtorkkola

shipit-source-id: b0436044d6