Fix MBR output path deletion for invalid paths
Summary:
Sometimes a ModernBuildRule can construct an output path which is not valid at the time that ModernBuildRule begins generating steps but is made valid by the steps generated by the Buildable's implementation. In this case, ModernBuildRule could generate a delete step for this invalid path, which would fail at runtime. This commit ensures that ModernBuildRule never issues delete steps for any path that is known to have elements in its path that are not directories, since they are guaranteed to fail at runtime.

This became apparent because A ModernBuildRule-ified Genrule emits an output path to a directory that is created by the Genrule's Buildable and not by MBR. In some cases, particularly when a build rule is changed between invocations, the output path observed by MBR is completely invalid. This is fixed by the Buildable's list of steps, but we need MBR to leave it alone.

Reviewed By: cjhopman

shipit-source-id: dcaee48fcd