Move writing RECORDED_PATHS to DefaultOnDiskBuildInfo
Summary:
Right now, we write the RECORDED_PATHS in the `BuildInfoRecorder`, but we write OUTPUT_SIZE, OUTPUT_HASH and RECORDED_PATH_HASHES in `OnDiskBuildInfo`.

Since we want to write all of these to a single file, we move the RECORDED_PATHS to be written in `OnDiskBuildInfo` too.

This should not be a behavioral change.

Reviewed By: asp2insp

shipit-source-id: fea95c308c