Replace string with enum for NDK ABI
Summary:
Use enum instead of strings for Android CPU-ABI pair.

Just to make code safer to change.

`NdkTargetArchAbi` gets another field, which is an ABI name used in Buck for compatibility with existing code. Actual string constants are copy-pasted from a switch in `NdkCxxPlatforms.java`, and that switch is now a switch by `NdkTargetArchAbi`.

Reviewed By: cjhopman

shipit-source-id: d9f2ff2dc3