Add support for apple_bundle
Summary:
Adds support for resource grouping to apple_bundle. This is implemented by using the APIs introduced before to filter the included resources for a particular apple_bundle.

We re-use all the pre-existing infrastructure built for link groups and just include the targets for the appropriate group, using the pre-existing semantics.

The grouping map is designed to be used together with link groups, as demonstrated in the integration test.

Reviewed By: blackm00n, paulqi1990

shipit-source-id: 988a5d9666