Add support for nullsafe_third_party_signatures option
Summary:
@public

Add a new buck config option under `infer_java` section that allows to specify a path (or a target)
to the directory with java method signatures for nullsafe to pick up.

Support for `--nullsafe-third-partysignatures` is a new feature in infer/nullsafe and it needs
support on the buck side.

Reviewed By: mityal

shipit-source-id: 0e3702af3e