Extend canary index id from 2 bytes to 7
Summary:
Change is to allow 2 3 byte indices with a separator (`%03d_%03d`), to support uniquely indexing split dex groups.

Uses 3 instead of 2 bytes for the per group index because we are already getting close to the maximum 2 byte limit, which could be hit in the case split dex groups are not used.

Use cases not using dex groups can format the index as `%07d`

Reviewed By: asp2insp

shipit-source-id: 5ef9f66230