Adds flexibility in collecting transitive targets
Summary: The old method, collectTransitiveBuildRules didn't return build rules. It returned constructor args. This updates the method to reflect that. It also adds a transform method for processing the TargetNodes that are traversed. It also adds a method for returning the TargetNodes themselves that will be used in a later change.

Reviewed By: asp2insp

shipit-source-id: 77ad32add2