Add caching for group map
Summary:
In practice, when using link groups, all targets will share the same group mapping (semantic, not reference equality). Consequently, we do not want to be recomputing the exact same mapping multiple times across the target group which can expensive in itself.

In this diff, we introduce a cache that avoids recomputation of the mapping for a particular graph.

shipit-source-id: b0f2d3dc19