Add a buck bin that is always stamped
Summary:
Add a way to always get a stamped buck binary. Normally this is not a concern (see notes in the BUCK file), but it can cause issues if one builds buck.pex locally, runs it in a repo with buck, makes more changes locally, rebuilds and tries to run again. The version doesn't change, so the old bin is used, and devs are confused because their new code isn't running.

Now you can always get a stamped binary with `buck build buck-local`

Reviewed By: gnuthor

shipit-source-id: 2e31841c6a