Add output attribute
Summary: Add an attribute that allows UDRs to have pre-declared artifacts based on a filename provided by users. This works similarly to the `out` parameter of `genrule`, and automatically declares the output artifacts for the user. This also allows our rules to fail properly if the rule implementation does not write to all of the outputs that a user expects.

Reviewed By: bobyangyf

shipit-source-id: 464a590eb5