Add output_list attribute
Summary:
Add an output_list attribute that allows us to create a list of outputs that should be used. These are then provided as a list of Artifacts to the user's implementation.

Note that right now this is less useful, as we limit things in the RAG to zero or one outputs when running in legacy/compatible mode, but it will have value down the line.

Reviewed By: bobyangyf

shipit-source-id: da17af4344