Change HasMultipleOutputs#getSourcePathToOutput(String) to return multiple SourcePaths
Summary:
Main changes are in:
* `HasMultipleOutputs`
* `AbstractSourcePathResolver`
* `DefaultSourcePathResolver`
* `IjProjectSourcePathResolver`
* `SourcePathResolverAdapter`
* `SourcePathResolverAdapterTest`

`SourcePathResolverAdapter` is a new class that converts the multiple paths to a single path. It does so by asserting that there is only one path in the paths returned from the resolver, and returns that one single path.

Note that `getAllAbsolutePaths()` does not have a check for `sourcepaths.size() == paths.size()` in the adapter because the collection of given source paths can map to same path

Reviewed By: bobyangyf

shipit-source-id: 680c2316e8