Add short-circuit for common case where we don't have runtime deps
Summary:
Now that `JavaLibrary` and descendents are `HasRuntimeDeps` we trigger the runtime deps pathway a lot more often.
This results in stacking up lots of instances of a no-op wrapper future (since the majority of libraries do not have runtime deps).
This diff adds a short-circuit to restore the original behavior in the absence of any declared `runtime_deps`

Reviewed By: IanChilds

shipit-source-id: e153cd8174