Use single ARTIFACT_METADATA file
Summary:
Instead of writing each bit of data to a separate file, we combine them into one file, called ARTIFACT_METADATA, which contains Json with each of the previous key names as the key into the map.

My understanding that Buck versioning means that I don't need to worry about extracting an artifact that has the old metadata format - since the buck version is part of the rule key?

Reviewed By: asp2insp

shipit-source-id: d8b1bbfd79