Extract Tool anonymous class to static class
Summary: Like D17831767 and the many other diffs that came before it, this diff extracts an anonymous inner class captured by a ModernBuildRule to a concrete, private static class. I was only aware that this was possible for Arg subclasses, but now it has been shown to be possible for Tool subclasses as well. Once this diff is in I'll do some static analysis on the codebase to try to identify other uses of this pattern and eliminate them.

Reviewed By: rajyengi

shipit-source-id: 7b5b44eb4c