Fix CellTest
Summary:
Previously test was written like:

```
BuildTargetFactory.newInstance(filesystem2.getRootPath(), "//does/not:matter");
```

and now it is:

```
BuildTargetFactory.newInstance(filesystem2.getRootPath(), "second//does/not:matter");
```

This is a bug in the test: cell root does not match cell name. Current implementation ignores it, but following diffs switch to cell names from cell paths start to fail in this test without this change.

Reviewed By: bobyangyf

shipit-source-id: 1e276f166b