Replace cellPath with cellFileSystem in UnflavoredBuildTargetView
Summary:
Cell path is no longer needed.

Previously cell path was used not just as a path, but also as a source of `FileSystem` object to construct `getBasePath`.

So to be able to continue constructing `getBasePath` `UnflavoredBuildTargetView` gets `cellFileSystem` object.

`cellFileSystem` is still included in hash and equals, so this diff does not solve the issue of hash consistency, more on this later.

Reviewed By: bobyangyf

shipit-source-id: 977b01b856