Fix ModernBuildRuleStrategyIntegrationTest in OSS test runs
Summary:
`trampoline.sh` has logic to find the right version of Java to execute based on the `BUCK_JAVA_VERSION` environment variable. That logic gets stripped out in the OSS version of the file, and we fall back to `java` on the path.

The problem is that `trampoline.sh` doesn't inherit the environment variables of its calling process (i.e. the thing that runs `GrpcRemoteExecutionServiceServer`). This is because `ActionRunner` relies on `DefaultProcessExecutor`, which strips existing environment variables if an environment is specified as part of the process creation args. This seems like a strange default choice, but I'm not inclined to change it.

This behavior causes us to use the wrong version of Java in `ModernBuildRuleStrategyIntegrationTest` in the OSS Java 11 tests, despite the right version of Java being at the front of the path in the server process, as `PATH` just gets stripped out.

The fix is to forward the value of `PATH` to the child process environment.

Reviewed By: rajyengi

shipit-source-id: b1ec1a2652