Pass correct Frameworks and PlugIns paths to swift-stdlib-tool
Summary:
`swift-stdlib-tool` is used to copy the Swift standard libraries to the app bundle. Even after ABI stability, this is still necessary for pre-12.2 iOSes—app thinning will handle the removal for more recent versions. The tool works by being passed a set of paths for binaries that contain Swift code. It will then determine the necessary libraries, and copy them.

For apps that contain either a framework or an app extension that use Swift, we pass additional directories. If, for example, the main app did not use Swift, but a framework did, and we did not do this, the app would crash on older iOSes due to the missing required Swift frameworks. This can be more subtle as well—if the framework uses a superset of the Swift libraries that the main binary does, it might appear that we're doing the right thing and copying, when in fact we're missing some frameworks. A similar concept applies to app extensions, except that only the extension will crash.

Currently, we are passing just the words "Frameworks" and "PlugIns" as arguments. This does not work, because `swift-stdlib-tool` expects an actual path. We should pass these relative to the bundle root, so that the tool can correctly scan the directories.

Reviewed By: apadalko

shipit-source-id: d921f324b0