Add new test to enforce MBR static class invariant
Summary: Given D18393909 and the endless stream of random bugfixes that fix the same root issue (non-static inner classes getting serialized by ModernBuildRule), this test attempts to locate all (or at least many) inner class subclasses of interfaces known to cross the ModernBuildRule boundary and asserts that they are static. Because Java reflection is only as good as the classpath in use, the new test target attempts to include as many Buck internals as its dependencies as possible so they are on the classpath to be reflected over.

Reviewed By: bobyangyf

shipit-source-id: 493ef423e9