Omit construction of SandboxProperties when unnecessary
Summary: An overwhelming majority of genrules run unsandboxed, which means that in the common case the SandboxProperties property of GenruleBuildable goes unused. This commit avoids constructing it entirely if it is known that it will not be used. SandboxProperties can be quite large, since it is a collection of four sets that include a large number of filepaths, particularly if the genrule has high in-degree.

Reviewed By: jtorkkola

shipit-source-id: a0ccafb85b