Allow using infer distribution folder as a provider of infer binary
Summary:
@public

Infer is distributed as a tarball with infer binary, required shared libs, and other
dependencies. Previously, we could only specify a **standalone binary** as a dependency in the infer
buck config, and then had to resort to out of bound delivery mechanisms to fetch the real binary
with its dependencies (e.g. dewey).

Now we can depend on the whole distribution folder and avoid using wrapper scripts for out of bound
fetching of infer.

Reviewed By: mityal

shipit-source-id: 4ed78cb45c