Allow apple_resource files copied to a bundle to be optionally codesigned
Summary:
Sometimes we want to copy a extra binary file into somewhere in a bundle (e.g. dylib to "Frameworks") and codesign it.  Right now Buck doesn't have a way of indicating individual files need to be codesigned (only other frameworks/bundles), so this adds a `codesign` flag to `apple_resource` to allow Buck to run codesigning on it.

The caller is responsible for ensuring the resource file is something which is signable.

Reviewed By: milend

shipit-source-id: 3b8758591b