Log execution events in machine readable log
Summary:
Adds events for both local and remote execution to the Buck machine log. This information allows us to do post-hoc analysis of actual rule execution since this information tells us exactly when Buck was running tools locally or scheduling things for remote execution.

In addition to Start and Finished events, this commit also emits RemoteExecutionScheduled events, which is useful to intuit when Buck scheduled an RE action and it has not yet started.

To accomodate the machine log, a number of new accessors and Json views were added to events, as well as some plumbing in the core executor to provide a BuildRule when firing an event and not just a BuildTarget. Since BuildTargets can be derived from BuildRules, this is a very minor change.

Reviewed By: rajyengi

shipit-source-id: 866e111044