Only count an exported dep as being used if it is not a first-order dep
Summary: Right now, if target A has a first order dep B which is not used but which exports C, and A needs C, then we don't mark B as unused (since if we remove it, A might stop compiling). However, if B is not B used and C is already a first-order dep of A, then it's reasonable to remove B. That is what this diff implements.

Reviewed By: asp2insp

shipit-source-id: 2411856889