Give a name to UnconfiguredTargetConfiguration
Summary:
`UnconfiguredTargetConfiguration` can now be addressed by `builtin//platform:unconfigured`.

Eventually `UnconfiguredTargetConfiguration` should go.

But while it exists, it's useful to be able to specify it in `--host-platform=`, in `--target-platforms=`, in `default_target_platform` or in platform detector. At least during migration or experimentations.

This is a simpler version of the diff where `builtin//platform:unconfigured` is not a built target.

Reviewed By: mzlee

shipit-source-id: 783aca1a7a