Order XZS metadata, expand dex validation in tests
Summary:
XZS files are unpacked in the order listed in the dex metadata. This happens to work because the `HashMap` that contains the metadata entries and the `ImmutableMultimap` that maps the output dexes to input dexes both have the same keys, and so have the same entry ordering.

This sorts them so it isn't broken by refactoring, and so it can be validated.

Adds dex/dex metadata validation utilities using `ZipInspector` and `DexInspector` to test this and the following dex sharding diff, as well as an integration test for XZS dex storage with multiple secondary dexes.

Reviewed By: asp2insp

shipit-source-id: 2144fa4ed4