Only use targets supplied at initialization in ProjectGenerator
Summary: There is a lot of redundant extra work here that was eliminated when we moved the focus matcher up a level. Before this, we are getting passed the targets we need to build, then going through the project graph node-by-node, and we are checking if it it's in this list. We can just simplify this considerably by using the raw list.

Reviewed By: williamtwilson

shipit-source-id: b24906136a