Process flavored targets before unflavored in ProjectGenerator
Summary:
The bug we are seeing is that in instances where a a bundle is included as a dep of the main bundle and a dep of a test, there will be two flavors in the target graph: one flavored, one unflavored. The order at which these appear seems to not be deterministic, so it's anyone's guess which one will be processed first. This is generally fine except in one specific instance: the schemas are awlays mapped by the flavored target; what this means is that if the unflavored target is processed first, the flavored will be skipped and there will be no entry in the graph for it; as a result, when a launch target is going to be added to the schema, there is no runnable built and thus it is left empty even though there is one.

To fix this, given that the order doesn't matter in the project generator, we can simply process the flavored targets first before the unflavored ones, which will fix this behavior as a side-effect.

shipit-source-id: 7a6c4dc869