Support per-library target_sdk_version
Summary:
This adds a new field to `apple_library`: `target_sdk_version`. The purpose of this field is to override the config-level minimum version. In large monorepos with multiple apps, a single version may not work, as some apps might wish to target a new iOS version. However, these apps will be unable to do so if another app wants an older version. With `-Wunguarded-availability` in Objective-C and always in Swift, using an API that is not available on the lowest API version without an availability check is a compiler error.

I used TDD to solve this. First, I wrote the test case added in `testdata`. I initially left off the `target_sdk_version` parameter as this would cause a parse error before the code to handle it was added. Originally, this would fail with this error:

```
stderr: Test.swift:6:5: error: 'TestFunction()' is only available in iOS 12.0 or newer
    TestFunction()
    ^
Test.swift:6:5: note: add 'if #available' version check
    TestFunction()
    ^
Test.swift:5:6: note: add available attribute to enclosing global function
func test() {
     ^
```

And then, a similar one for Objective-C. After this change, and the addition of the `target_sdk_version` parameter in the `apple_library` test rules, it passes.

To solve this for Swift, I've replaced the plain-old strings for target triples with a higher-level `SwiftTargetTriple` type. This allows us to replace *just* the version when a library overrides it, keeping the other attributes of the triple from the underlying platform. Thus, I've updated some other code and tests to use this instead of a string.

Reviewed By: williamtwilson

shipit-source-id: 8f0f76ae14