Remove benchmark binary target
Summary:
This target results in 40M binary in `buck-out`.

It's not scalable do it for each benchmark.

Unfortunaly, Buck doesn't provide a Java rule which exposes a java program but does not create a fat jar on build (like `java_test` but without junit). So just remove the binary. Benchmark can be still executed from Idea.

Reviewed By: bobyangyf

shipit-source-id: 0370d174c5