MergedTargetNode and MergedTargetGraph utilities
Summary:
`MergedTargetNode` is a set of `TargetNode` objects with the same unflavored build target.

`MergedTargetGraph` is a version of `TargetGraph` where nodes are `MergedTargetNode` and edges are all edges from underlying `TargetGraph`. E. g. for four nodes `TargetGraph`:

```
a -> c(//:p)
b -> c(//:q)
```

`MergedTargetGraph` is:

```
a -> c
b -> c
```

This is a groundwork for allowing multiple configurations for the same target on one build.

Reviewed By: bobyangyf

shipit-source-id: c0bb44d022