Switch to MergedTargetNode in AbstractQueryCommand
Summary:
Currently it's not allowed to have multiple targets with the same unconfigured target but with different configurations in the same graph. It is explicitly prohibited because output paths in `buck-out` clash.

We will soon remove this restriction (will generate unique paths for different configurations).

When we do so, we would need to do something about `buck query`: currently it outputs the configured graph, but outputs as a map with unconfigured unflavored target as a key.

The plan is to merge all configurations in `buck query` output (concatenating lists attributes and choosing arbitrary values for non-concatenatable attributes).

This diff is a ground work for this: `MergedTargetNode` is unconfigured. Currently `buck query` outputs first node from underlying nodes. Later it will merge them. The merge is not implemented yet (it's not possible to implement merge yet, because configuration switching is not landed yet).

Also, we will need to implement `cquery` eventually without attribute merge.

Reviewed By: bobyangyf

shipit-source-id: 28ac7bcb87